🚅 perfs: move payments judge from psbtUtil to payments
