+++ src/payments/index.d.ts
@@ -46,3 +46,10 @@ export type StackElement = Buffer | number;
 export type Stack = StackElement[];
 export type StackFunction = () => Stack;
 export { embed, p2ms, p2pk, p2pkh, p2sh, p2wpkh, p2wsh, p2tr };
+export declare const isP2MS: (script: Buffer) => boolean;
+export declare const isP2PK: (script: Buffer) => boolean;
+export declare const isP2PKH: (script: Buffer) => boolean;
+export declare const isP2WPKH: (script: Buffer) => boolean;
+export declare const isP2WSHScript: (script: Buffer) => boolean;
+export declare const isP2SHScript: (script: Buffer) => boolean;
+export declare const isP2TR: (script: Buffer) => boolean;
