+++ src/psbt/input/script.d.ts
@@ -1,6 +1,6 @@
 /// <reference types="node" />
 import { PsbtInput } from "bip174/src/lib/interfaces";
-import { PsbtCache, ScriptType } from "../interfaces";
+import { GetScriptReturn, PsbtCache, ScriptType } from "../interfaces";
 export declare function getMeaningfulScript(script: Buffer, index: number, ioType: 'input' | 'output', redeemScript?: Buffer, witnessScript?: Buffer): {
     meaningfulScript: Buffer;
     type: 'p2sh' | 'p2wsh' | 'p2sh-p2wsh' | 'raw';
@@ -14,3 +14,4 @@ export declare function getScriptAndAmountFromUtxo(inputIndex: number, input: Ps
     script: Buffer;
     value: number;
 };
+export declare function getScriptFromInput(inputIndex: number, input: PsbtInput, cache: PsbtCache): GetScriptReturn;
