+++ ts_src/psbt/input/index.ts
@@ -1,10 +1,11 @@
 import { PsbtInput } from "bip174/src/lib/interfaces";
 import { checkTaprootInputForSigs, isTaprootInput } from "../bip371";
-import { checkInputForSig, isP2TR, pubkeyInScript } from "../psbtutils";
+import { checkInputForSig, pubkeyInScript } from "../psbtutils";
 import { Output, Transaction } from "../../transaction";
 import { PsbtCache } from "../interfaces";
 import { checkTxInputCache, nonWitnessUtxoTxFromCache } from "../global/cache";
 import { getMeaningfulScript, getScriptAndAmountFromUtxo, getScriptFromUtxo, scriptWitnessToWitnessStack } from "./script";
+import { isP2TR } from "../../payments";
 
 export function checkInputsForPartialSig(inputs: PsbtInput[], action: string): void {
   inputs.forEach(input => {
